refactor(ui-demo): Implemented new project header into demo app
