+++ frontend/alanda/apps/ui-demo/src/app/components/projects-controller/projects-controller.component.html
@@ -1,58 +1,3 @@
 <ng-container *ngIf="state.select() | async as s">
-  <alanda-page-header
-    [project]="s.project"
-    [activeTabIndex]="activeTabIndex"
-    [activePhase]="'active'"
-    [tabs]="s.tabs"
-    [user]="s.user"
-    (activeTabIndexChange)="activeTabIndex = $event"
-  >
-    <ng-container [ngSwitch]="activeTabIndex">
-      <div *ngSwitchCase="'0'">
-        <alanda-project-overview [project]="s.project" [user]="s.user">
-          <div additional>
-            <alanda-vacation-project-properties
-              [project]="s.project"
-              [user]="s.user"
-            ></alanda-vacation-project-properties>
-          </div>
-        </alanda-project-overview>
-      </div>
-      <div *ngSwitchCase="'1'">
-        <alanda-pio-component [pid]="s.pid"></alanda-pio-component>
-      </div>
-      <div *ngSwitchCase="'2'">
-        <alanda-history-grid-component
-          *ngIf="s.project"
-          [projectGuid]="s.project.guid"
-        ></alanda-history-grid-component>
-      </div>
-      <div *ngSwitchCase="'3'">
-        3
-      </div>
-      <div *ngSwitchCase="'4'">
-        4
-      </div>
-      <div *ngSwitchCase="'5'">
-        5
-      </div>
-      <div *ngSwitchCase="'6'">
-        6
-      </div>
-      <div *ngSwitchCase="'7'">
-        7
-      </div>
-      <div *ngSwitchCase="'8'">
-        8
-      </div>
-      <div *ngSwitchCase="'9'">
-        9
-      </div>
-      <div *ngSwitchCase="'10'">
-        10
-      </div>
-    </ng-container>
-  </alanda-page-header>
-
   <router-outlet></router-outlet>
 </ng-container>
