+++ frontend/alanda/apps/ui-demo/src/app/components/projects-controller/projects-controller.component.ts
@@ -3,19 +3,12 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap, tap, map, distinctUntilChanged } from 'rxjs/operators';
 import { Subscription } from 'rxjs';
 import { RxState } from '@rx-angular/state';
-import {
-  AlandaPageTab,
-  AlandaProject,
-  AlandaProjectApiService,
-  AlandaUser,
-  BaseState,
-} from '@alanda/common';
+import { AlandaProjectApiService, AlandaUser, BaseState } from '@alanda/common';
 import { UserStoreImpl } from '../../store/user/user.store';
 
 export interface ProjectControllerState extends BaseState {
   pid: string;
   user: AlandaUser;
-  tabs: AlandaPageTab[];
 }
 
 @Component({
@@ -23,48 +16,7 @@ export interface ProjectControllerState extends BaseState {
   providers: [RxState],
 })
 export class ProjectsControllerComponent {
-  activeTabIndex = 2;
   paramSub: Subscription;
-
-  activePhase: string;
-
-  overviewTab: AlandaPageTab = {
-    name: 'Overview',
-    icon: 'pi pi-bookmark',
-    phase: null,
-  };
-
-  pioTab: AlandaPageTab = {
-    name: 'Pio',
-    icon: 'pi pi-eye',
-    phase: null,
-  };
-
-  historyLogTab: AlandaPageTab = {
-    name: 'History Log',
-    icon: 'pi pi-clock',
-    phase: null,
-  };
-
-  peopleTab: AlandaPageTab = {
-    name: 'People',
-    icon: 'pi pi-users',
-    phase: null,
-  };
-
-  // routerParams$ = this.router.events
-  // .pipe(
-  //   filter((event: RouterEvent): boolean => (event instanceof NavigationEnd)),
-  //   map(() => this.router.routerState.snapshot.root),
-  //   // @TODO if we get away from global task managing dete this line and move coed
-  //   map(snapshot => this.collectParams(snapshot)),
-  //   tap(snapshot => console.log('sn', snapshot)),
-  //   distinctUntilChanged(sn => sn.projectId),
-  //   switchMap((params) => {
-  //     console.log(params);
-  //     return this.projectService.getProjectByProjectId(params.projectId);
-  //   })
-  // );
 
   routerParams$ = this.route.params.pipe(
     map((p) => p.projectId),
@@ -77,46 +29,12 @@ export class ProjectsControllerComponent {
     }),
   );
 
-  getPidFromProject$ = this.state.select('project').pipe(
-    map((project: AlandaProject) => {
-      return project.processes.find((proc) => proc.relation === 'MAIN')
-        .processInstanceId;
-    }),
-  );
-
   forwardByType$ = this.state.select('project').pipe(
-    tap((project: AlandaProject) => {
+    tap((project) => {
       this.router.navigate([project.projectTypeIdName.toLowerCase()], {
         relativeTo: this.route,
         skipLocationChange: true,
       });
-    }),
-  );
-
-  tabs$ = this.state.select('project').pipe(
-    switchMap((project) =>
-      this.projectApiService.getPhasesForProject(project.guid),
-    ),
-    map((response) => {
-      const phases = response.filter(
-        (phase) => phase.pmcProjectPhaseDefinition.idName !== 'PHASE1',
-      );
-
-      const phaseTabs: AlandaPageTab[] = phases.map((phase) => {
-        return {
-          name: phase.pmcProjectPhaseDefinition.displayName,
-          icon: 'pi pi-play',
-          phase: phase,
-        };
-      });
-
-      return [
-        this.overviewTab,
-        this.pioTab,
-        this.historyLogTab,
-        ...phaseTabs,
-        this.peopleTab,
-      ];
     }),
   );
 
@@ -129,13 +47,7 @@ export class ProjectsControllerComponent {
     private userStore: UserStoreImpl,
   ) {
     this.state.connect('project', this.fetchProjectByProjectId$);
-    this.state.connect('pid', this.getPidFromProject$);
     this.state.connect('user', this.userStore.currentUser$);
-    this.state.connect('tabs', this.tabs$);
     this.state.hold(this.forwardByType$);
-  }
-
-  phaseChange(event) {
-    this.activePhase = event;
   }
 }
