+++ frontend/alanda/apps/ui-demo/src/app/features/vacation/components/vacation-project-details/vacation-project-details.component.html
@@ -1,3 +1,7 @@
+<alanda-vacation-page-header
+  [project]="project"
+  [user]="user"
+></alanda-vacation-page-header>
 <alanda-project-and-processes
   [user]="user"
   [project]="project"
