+++ frontend/alanda/libs/common/src/lib/form/alanda-task-form.service.ts
@@ -27,14 +27,12 @@ import { AlandaTaskApiService } from '../api/taskApi.service';
 import { MessageService } from 'primeng/api';
 import { AlandaTitleService } from '../services/title.service';
 import { AppSettings, APP_CONFIG } from '../models/appSettings';
-import { AlandaPageTab } from '../components/page-header/page-header.component';
 
 export interface AlandaTaskFormState {
   task?: AlandaTask;
   project?: AlandaProject;
   loading?: number;
   closeAfterComplete?: boolean;
-  tabs?: AlandaPageTab[];
   //  rootFormData: { [controlName: string]: any }
 }
